+++ internal/fetcher/http_fetcher.go
@@ -18,6 +18,7 @@ const (
 )
 
 type Http struct {
+	httpClient *http.Client
 	limiter    *rate.Limiter
 	reqTimeout time.Duration
 }
@@ -26,6 +27,9 @@ type Http struct {
 func NewFetcher(httpConf config.HttpClient) *Http {
 	limiter := rate.NewLimiter(rate.Every(time.Minute/time.Duration(httpConf.RatePerMinute)), 1)
 	return &Http{
+		httpClient: &http.Client{
+			Timeout: httpConf.TimeoutMS,
+		},
 		limiter:    limiter,
 		reqTimeout: httpConf.TimeoutMS,
 	}
@@ -47,9 +51,9 @@ func (f *Http) Fetch(ctx context.Context, ipfsID string) (json.RawMessage, error
 	req.Header.Set("Content-Type", "application/json")
 	req.Header.Set("User-Agent", "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36")
 
-	resp, err := http.DefaultClient.Do(req)
+	resp, err := f.httpClient.Do(req)
 	if err != nil {
-		return nil, fmt.Errorf("call ipfs: %w", err)
+		return nil, fmt.Errorf("call ipfs %s: %w", ipfsID, err)
 	}
 
 	defer resp.Body.Close()
@@ -61,7 +65,7 @@ func (f *Http) Fetch(ctx context.Context, ipfsID string) (json.RawMessage, error
 
 	isSuccessStatus := resp.StatusCode >= 200 && resp.StatusCode < 300
 	if !isSuccessStatus {
-		return nil, fmt.Errorf("http status code: %d, body: %s", resp.StatusCode, body)
+		return nil, fmt.Errorf("ipfs: %s, http status code: %d, body: %s", ipfsID, resp.StatusCode, body)
 	}
 
 	isJsonValid := json.Valid(body)